BL: move file to folder
