+++ welegend/app/(pages)/introduce/_components/Lectures.tsx
@@ -7,14 +7,13 @@ import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { Button, Carousel } from 'antd';
 import { chunk } from 'lodash';
 import { useEffect, useRef, useState } from 'react';
-import { CardInfo } from './CardInfo';
+import { CardInfo } from './cards';
 
 export function Lectures() {
   const [listItems, setListItems] = useState([
     0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10,
   ]);
   const chunks = chunk(listItems, 4);
-
 
   const [numChunks, setNumChunks] = useState(4);
