+++ welegend/app/(pages)/introduce/_components/cards/CardInfo.tsx
