+++ welegend/app/(pages)/introduce/_components/cards/ValueInfo.tsx
