+++ welegend/app/(pages)/introduce/_components/cards/index.ts
@@ -1,0 +1,2 @@
+export * from './CardInfo';
+export * from './ValueInfo';
