+++ welegend/app/(pages)/introduce/_components/index.ts
@@ -1,4 +1,5 @@
 export * from './BannerHeader';
+export * from './Buildings';
 export * from './Lectures';
 export * from './Thanks';
 export * from './TruthStory';
